Update de_de.json
updated the "sync_status" to be more german. I'm still not 100% happy but I used mostly non-autotranslated and bridged it heavily on being grammatically correct instead of "true to the word"
I also replaced the breastcancer awareness with a german one. I think it makes more sense that way. But it'd also need a link to the german website. So potentially a need for "wildfire_gender.cancer_awareness.url"? I don't think it makes sense for different translations to go to one-stop resource.