Various small UI user experience fixes
- Remove/rewrite a bunch of redundant tooltips in character settings

Tooltips such as "Enables Breast Phyiscs" are bad, as they just repeat
the setting name, but worded slightly differently.

This changes them to actually describe the setting in better detail, or
removes the tooltip entirely if the setting name is already fairly
self-explanatory.

- Don't re-open the main player settings menu when the gender is updated

This is aimed at avoiding the narrator repeating the screen title
whenever this setting is changed, as it logically wouldn't make sense to
do so in this case.

- Move the playing with creator prompt down to not overlap with the GUI

Also moves it under the breast cancer awareness month banner when
relevant.