Copy player breast settings to armor stands
This splits rendering into two different renderers, one for the base
skin, and another for the armor layer, which brings rendering more
in-line with how vanilla handles such rendering.

This additionally splits GenderPlayer into two classes, a new
EntityConfig, and the now renamed PlayerConfig, which subclasses the
former.

The way that this attaches the renderer to the armor stand is that when
a player equips a chestplate onto an armor stand, the mod will copy
their settings onto the chestplate item stack in the `WildfireGender`
NBT tag, and read that on the client-side for rendering.

This shouldn't cause any issues with modded armors, as it will only do
so when the armor config specifies to do so, which by default will only
apply on armors with complete physics resistance (with gold added as a
manual exception).