Add client configuration options
This is meant to provide some parity between the Forge version and
Fabric, although this doesn't provide perfect parity, as this doesn't
load the Forge config file due to Forge using TOML for configuration
files.

This also adds a GUI to modify these values in-game, without having to
manually modify the config file.

The armor physics override setting has also been moved to this client
configuration, to match the intended use case of the setting.