improve sidedness for mod initializer
this also changes hurt sounds to be purely client-sided to work around
fabric api doing some relatively strict registry syncing, which I didn't
consider when initially choosing to make this occur on both the client-
and server-side.

this won't fully solve the issue in singleplayer worlds opened to LAN,
but there's not a whole lot we can do here outside of forcing fabric-api
to simply not sync our sound registry entry with an ugly bodge mixin.