switch to using '$mod+$minecraft' versioning instead of '$minecraft-$mod'
the way this was done previously led fabric to infer that the mod
version was the minecraft version, and not the (intended) other way
around.

this also provides some level of parity with the neoforge version
change made in aa55cbe5284b26bc7d62d7f9111bf7f9ccb5757d, which simply
drops the minecraft version from the mod version string entirely.